requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
aiosql==10.4
aiosqlite==0.20.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.7.4
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
fancycompleter==0.9.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
linkify-it-py==2.0.3
markdown-it-py==3.0.0
mdit-py-plugins==0.4.1
mdurl==0.1.2
msgpack==1.0.8
multidict==6.0.5
ollama==0.3.1
packaging==24.1
pdbpp==0.10.3
pillow==10.4.0
pluggy==1.5.0
pygments==2.18.0
pyperclip==1.9.0
pyreadline==2.1
pyrepl==0.9.0
pytest==8.3.2
pytest-asyncio==0.23.8
python-dotenv==1.0.1
rich==13.7.1
rich-pixels==3.0.1
ruff==0.5.7
shellingham==1.5.4
sniffio==1.3.1
textual==0.76.0
textual-dev==1.5.1
tomli==2.0.1
typer==0.12.3
typing-extensions==4.12.2
uc-micro-py==1.0.3
wmctrl==0.5
yarl==1.9.4
